# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/workers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Get workers",
                "operationId": "many-workers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Worker name (pattern)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Worker system (pattern)",
                        "name": "system",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "Worker status (possible values)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "Worker type (possible values)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of workers",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.WorkerShort"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Create new worker",
                "operationId": "create-worker",
                "parameters": [
                    {
                        "description": "New worker entry",
                        "name": "worker",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.WorkerInput"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/workers/{worker_name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Get the single worker",
                "operationId": "single-worker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Worker name",
                        "name": "worker_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Requested worker",
                        "schema": {
                            "$ref": "#/definitions/model.Worker"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Update worker",
                "operationId": "update-worker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Worker name",
                        "name": "worker_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated worker entry",
                        "name": "worker",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.WorkerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated worker",
                        "schema": {
                            "$ref": "#/definitions/model.Worker"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "workers"
                ],
                "summary": "Delete worker",
                "operationId": "delete-worker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Worker name",
                        "name": "worker_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Force deletion",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Build": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.BuildStatus"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildStep"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "worker_name": {
                    "$ref": "#/definitions/sql.NullString"
                }
            }
        },
        "model.BuildLog": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "idx": {
                    "type": "integer"
                },
                "output": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.BuildStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "BuildScheduled",
                "BuildRunning",
                "BuildSuccessful",
                "BuildFailed",
                "BuildCanceled"
            ]
        },
        "model.BuildStep": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "$ref": "#/definitions/time.Duration"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BuildLog"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Worker": {
            "type": "object",
            "properties": {
                "active_builds": {
                    "type": "integer"
                },
                "address": {
                    "type": "string"
                },
                "builds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Build"
                    }
                },
                "capacity": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.WorkerStatus"
                },
                "strategy": {
                    "type": "integer"
                },
                "system": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.WorkerType"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.WorkerInput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "private_key": {
                    "type": "string"
                },
                "strategy": {
                    "type": "integer"
                },
                "system": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.WorkerType"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.WorkerShort": {
            "type": "object",
            "properties": {
                "active_builds": {
                    "type": "integer"
                },
                "address": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.WorkerStatus"
                },
                "strategy": {
                    "type": "integer"
                },
                "system": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.WorkerType"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.WorkerStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "WorkerIdle",
                "WorkerUsed",
                "WorkerUnreachable"
            ]
        },
        "model.WorkerType": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "WorkerStatic",
                "WorkerDockerHost"
            ]
        },
        "sql.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
        "time.Duration": {
            "type": "integer",
            "enum": [
                1,
                1000,
                1000000,
                1000000000
            ],
            "x-enum-varnames": [
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second"
            ]
        }
    }
}